Fix heise.de main RSS feed
- h1 title is redundant with title in <head>
- always request ?seite=all to get multipage articles
- the regular <img src> element is hidden under noscript, so unhide that
- ... and drop a-img and other elements that can't work without CSS

We also need to strip the a-lightbox tags (but not their content). While
the web frontend as well as the overview page in the Android app would
show the <img> tag inside just fine, as soon as you read the article in
the Android app, the <img> inside <a-lightbox> would no longer be
rendered. ¯\_(ツ)_/¯