Fix the heise.de recipe and make Bilderstrecken show the first image.
Turns out, the feed always has something after the .html url, so the
match didn't always stick. This now should force ?seite=all more
consistently.

Also, drop the <a-bilderstrecke> tags which would hide the <img src>
tags within from being displayed, especially on the Android app.

Unlike the arstechnica hack, this does not show all images, but at least
the first one plus caption is visible and the user can decide to open
the full article in the browser or not.